# Check cached postings TTL before returning from cache + metrics